azure-cognitive-services/requirements.txt
python-dotenv==1.0.1
pillow==11.3.0